Run render passes in the correct order
Previously we were using render passes in whatever order they happened to appear in the ShaderToyDocument, after swapping the "image" pass to the end. This isn't robust: even if the ShaderToy API always returns them in the correct order, there's no guarantee that externally authored JSON files will have them in the correct order.

We now rely on the passes having names consistent with ShaderToy and use that to sort & filter them into the correct order.

The way we swapped the image pass to the end was buggy because the assetIDToRenderPass mapping wasn't being updated when we did the swap. It's kind of a miracle that it worked at all for multipass renders. We no longer do the swapping - it's no longer needed -
 which fixes this bug.

Previously we were assuming that buffer passes could run in any order because they should always read from the "front buffer" output (i.e. the values calculated for the buffer on the *previous* frame). This is clearly wrong: if you had render passes in a chain, each writing out the value of the previous pass, there would be up to a 5 frame delay for values to move all the way through from Buf A to the Image pass. So what we do now is read from the back buffer if we're reading from a render pass that has already executed in the current frame and only fall back to using the front buffer if the source pass has not yet run in the current frame.

We also now clear all output textures before rendering frame 0, or immediately after resizing them. This is to ensure that render passes cannot end up reading garbage values from them.