Don't crash when textures are missing
If the input file specifies a texture or cubemap that doesn't exist, or can't be loaded for some reason, we now fall back to using the 1x1 all-black texture instead.

This is currently buggy for cubemaps, because the placeholder texture is a 2D image rather than a cubemap, so we'll generate an incorrect sampler type for the shader.