Look up renderpasses using type and index instead of name
Can't use pass names because the ShaderToy API doesn't return consistent names for the passes (e.g. some shaders have "Buf A", others have "Buffer A"). 

Note that this will return incorrect results for buffers if you have e.g. Buf B but not Buf A (in this case, the lookup for Buf A would end up returning Buf B instead of an empty renderpass as you'd expect).