Fix nvidia-specific error in KifsOctahedrons.json
The shader for Buf A had a `clamp` call with the value to be clamped as the last parameter (wrong) rather than the first (correct). This gives undefined behaviour according to the GLSL spec; the result was that the shader was broken on Nvidia but worked fine on Intel and in-browser.