Fix keyboard handling, toggle keyboard/mouse shader input on and off
Fixed some bugs in the keyboard handling which made it incompatible with ShaderToy. It *should* be correct now... I think.

Added the ability to enable/disable sending keyboard events to the shaders. Needed this because otherwise we get keyboard shortcuts being triggered as well (e.g. pressing the space bar will send that keypress to the shader, but also pause the shader). You can always toggle keyboard shader input on and off using F2, whichever mode you're in.

Did the same for mouse input to the shader as well. You can toggle that on and off using F3. This lays the groundwork for being able to pan and zoom the shader output.