Add Save, Save As, Extract GLSL and Inline GLSL menu items
Extract GLSL saves the code for each renderpass, including the common pass, into an external GLSL file and updates the renderpass with a reference to the file. The filename is the name of the shader toy and the name of the renderpass, separated by a hyphen and with any spaces removed.

Inline GLSL does the reverse: for renderpasses which have an external file reference it loads the file and sets their `code` field to the file contents, then removes the reference to the external file.

Save and Save As do what you'd expect.