Download a (hardcoded) shader from ShaderToy
Added code using QNetworkAccessManager to download a shader from ShaderToy. This is just me getting my head around the Qt's networking classes, it's not usable yet.

The code downloads a shader with a hardcoded ID from ShaderToy. and saves it in a hardcoded file. It doesn't yet download any referenced assets, check if we've already downloaded a copy of it, or anything like that.