Add support for panning and zooming the output image
Added a simple-but-flexible system for mouse bindings, along similar lines to the key binding system.

Removed the option to toggle between sending all mouse input to the shader or using it for the application. There's no need for it now, this is handled through the mouse binding system just like everything else. From a user point of view: left mouse with no modifiers goes to the shader, anything else is handled by the application.