Add proof of concept for video input
Work in progress on supporting video inputs to shaders.

Currently there is a single hard-coded video that is always loaded and played back, whether or not the shader has any video inputs. If there are any render passes with video inputs, then we use that same video for all of them.

The tricky part was getting the video to stream into an OpenGL texture. Qt's API is designed to *push* video frames, but I wanted to *pull* them instead so that I could ensure we're never trying to update a texture while rendering with it. What I ended up with was a custom QAbstractVideoSurface that keeps a copy of the most recently presented QVideoFrame; and I upload it to an OpenGL texture once at the start of each iteration of the render loop.

This is not yet as efficient as it could be: if the video pixels are already on the GPU, they will be round-tripped via the CPU. I tried to fix this by adding special-case handling for the video frames with `handleType() == GLTextureHandle`, but couldn't coax Qt on Windows into using GLTextureHandle (it always seems to use NoHandle).

VERY IMPORTANT: Qt's video support on Windows uses whatever DirectMedia codecs you have installed. If you haven't installed any additional codecs manually you won't be able to play back Ogg video. See the README for more info about this and where to find a set of codecs that includes Ogg support.