Support multiple video inputs
This is gradually moving closer to proper video support.

We now load the videos specified in the shader JSON, instead of a hard-coded video. If there is more than one video input, that will be handled correctly too. 

There is a currently a hardcoded limit of 4 different videos, but this should be ok for now since ShaderToy only has 4 different videos available. I'll remove this limit in a future commit.

There are no widgets (yet) to control playback of individual videos.

All videos are muted for now & there is no way for a user to unmute them.

Added another standard video asset path to the file cache and added the NeuralNetVideSample example file.