Fix bug where video keeps updating for a while after you've paused playback
This was happening because QMediaPlayer was presenting one or more new frames in the time it took for the pause() or stop() signal to be processed. We stop the update loop before the last of those frames arrives, so the new frames aren't shown; but when something forces an update (such as popping a menu up over the RenderWidget), the video visibly jumps to the new frame.

I've worked around this by adding a "paused" state to the TextureVideoSurface. While paused it will just discard any new frames that get passed to its present() method. Pausing and unpausing the TextureVideoSurface at the appropriate times (the changes in RenderWidget.cpp) fixes the issue.