Add support for the Cube A render pass
Refactored the GLSL template so that the code for the various main functions is all inline & selected using a #define, rather than passing in the main function as a macro.

When setting up the render passes, we now check whether an input is a render pass by using the new `inputIsRenderPass()` method (which tests against the input's ID) instead of checking whether `input.ctype == "buffer"`. This means that renderpasses of types other than buffer are handled correctly too.