Upload audio data to a texture
Added the TextureAudioSurface class, which works similarly to TextureVideoSurface. It holds a copy of the most recently supplied audio buffer and has the logic for uploading it to a texture when required.

This is only half of what is required by audio inputs: we upload the sample data from the audio (row 1 of the texture), but we're not yet calculating or uploading the frequency data (row 0 of the texture). At the moment we simply copy the data from row 1 into row 2 as a placeholder

This does not work with Qt 5.11 or 5.12 on Windows, due to a bug which broke their QAudioProbe support. It does work with the current Qt 5.13 alpha though (the bug was marked as fixed in 5.13) and *should* work on Qt 5.10 & earlier.