Display cubemaps as a cross of images
When you're viewing the output of the Cube A pass, or when a thumbnail for the Cube A pass is showing, we now display it as a cross of images.

Previously we just attempting to display the cubemap texture as a 2D texture, which caused a lot of OpenGL errors.

This closes #37 (thumbnails for cubemaps) and closes #40 (viewing the cubemap pass).